feat: moving graphs for accel/gyro
This commits implements and enables a moving chart for plotting
accelerometer and gyroscope data. It plots data from the last 10
seconds.

Additionally, this commit makes slight UI changes and updates the web
manifest with the proper app information.

As usual, these commits are a little bloated and nonsensical but we're
building quickly. Sorry!