feat: stop parsing BLE notify data when hidden
When the app is hidden the browser may throttle the tab, leading to
a buildup of messages that cause the app to hang when focus is restored.
To prevent this, this commit uses the Page Visibility API to detect when
the app/page is not visible and stops parsing data. Upon return, a
notification is also shown to the user to notify them that data/warnings
may be missing.

More info about the page visibility API:
https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API