feat: in-app notifications
This commit extends the previous work in notifying the user that data is
not refreshed when the app is hidden and allows for the creation of an
infinite number of simple notifications. Because I'm not sure what to
use as a key, I am using the time since epoch in milliseconds.

The notification color, text and display duration can be customized by
passing different arguments to showNotification.