autodeploy: build from main branch
Currently Github actions builds the web app and pushes it to Github
pages each time a push is made to the rover-web branch. This is because
I was unsure whether we were going to build a web app or a native api so
I started tinkering on the rover-web branch.

Since I am most likely working on a web app for this project, I am going
to pull rover-web into main and start making branches for bugfixes +
feats like a more typical workflow. Or maybe I'll just work on main.

Regardless, this commit prepares for the pull to main by changing the
Github action workflow in ghpages.yml to track/build from main.