feat: switch between dark and light modes
While dark mode is the default, a toggle has been added in the settings
tab to switch to light mode. This commit breaks out the Settings tab
into a component called TabSettings and implements local-storage in
order to save preferences.

Also changed page visibility listener to a variable from a React state
variable. Hopefully this prevents React from doing state updates as the
page is unloading.