feat: capture keyboard and send in queue
To send keyboard data to the rover, the app will listen to key down and
key up events to listen for wasd. Upon recipt of these key events, new
functions within Rover.js will calculate the proper checksum and queue a
GATT message for transmission. A queue was implemented because one is
not built in to the Web Bluetooth API leading to collisions.

Keyboard input is only sent when the app is connected to the rover and
the rover is in 0x03 mode, which can not be toggled using a new entry in
the drive tab.