fix: remove event listeners correctly
Redefined anonymous functions passed to event listeners as class
variables using arrow notation so removing event listeners work
correctly. This should reduce errors with disconnecting/connecting from
the rover repeatedly. As well, when the page is invisible, the TX
listener handler is removed from the TX characteristic, hopefully
preventing a buildup of notifications and lag if the rover is connected
but the tab is in the background for a long time (to be tested on mobile
after this build/deploy completes.)

Finally, this commit introduces a 75ms delay between sending messages to
the RX characteristic. This allows the rover time to process the last
sent message if messages are sent in rapid succession (ie. keyboard
control.)