feat: set target speed
The targeted speed of the rover can be set while the rover is in mode
2 using buttons in the drive tab or using the up/down arrow keys on the
keyboard. This commit implements this behaviour, sending the command
described in the communication format document corresponding to
increasing/decreasing/setting a targeted speed value.

The rover speed displayed by the UI is updated when the rover reports
it's current targeted speed as part of it's TX status messages.
This means there is a slight delay between pressing the button/key to
the UI reflecting the speed increase/decrease. This is intentional as
this helps prevent the UI and the rover states from getting out of sync
without implementing more complex state tracking functions.

This commit also adds images showing what keyboard buttons correspond
with what action and tweaks the UI to display better across different
screen sizes.