feat: add D-Pad and controller state table
This commit adds a D-Pad that can be used by both mouse and touch
inputs. Additionally, a table has been added that shows the status of
each controller as well as some of the values that are planned to be
reported. Right now this table, as well as other parts of controller
state, are not dynamic. This will be implemented as we get closer to
making the rover move.
Finally, motor control mode can be exited by pressing the esc key and
decrease/increase speed has been changed to the left and right arrow
keys to better match the range input. Images have been updated to
reflect this.