arduino: inital commit
This commit adds the work currently done for the rover microcontroller.
All work has currently been done on a Adafruit Feather nRF52832,
generating simulated data when needed. The code has been structured to
make implementation on the actual board as easy as possible when it
arrives. This will involve writing a new "board driver" (in quotes
because it is not an actual driver) that implements the function
prototypes outlined in roverdriverarduino.h.

Currently the code allows data to be sent to and from the controller
using the Nordic UART interface. Mode switching, control and device
telemetry is also functional (although IMU data is random).