feat: logging for IMU and system stats
User-initiated logging of rover data has been added in this commit using
JsStore to interface with indexeddb, local storage managed by the
browser.

Using the log tab, the user can pick what characteristics to log and
start/stop logging. The logging interval is set to 2 Hz. Previous logs
are also displayed in the same tab and can be exported as CSV for
further analysis and processing.

Each log file is represented using a table in indexeddb, with a main
table recording the names of the "files" as well as their creation date.
The table name for each file is randomly generated as a UUID v4 and is
normally hidden, however a setting has been added to show this for
easier debugging.