feat: show prompt if serviceworker update ready
When there is a serviceworker waiting to be installed, send an event
that is caught by the App component in componentDidMount. The app will
then display a in-app notification with a button to install and reload
the page.

The in-app notification component has been updated to allow for an
action button and associated click handler to be passed via props.