fix: pause/restart logging on visibility change
This commit corrects the behaviour of logging when the visibility of the
app changes. The app will not stop the logging timer while the user is
away and restart it using when the app is visible again. It does this by
storing the parameters of the log (ex. table name, interval, what to
log) in the logging state variable of App.