app: change imu format to Float32
Instead of transmitting imu data as a string, it is faster and more
conservative to transmit it as a Float32. Instead of doing string
manipulation to extract each axis, the addMovingData function not checks
that the byte array being parsed has a length of 13 (1 byte subject,
three 4 byte axis). It then uses a DataView to extract each axis as a
Float32.

This commit also changes the unit of the magnetometer to uT and
increases the data shown in charts to 10 seconds at 4 Hz, for 40 total
points plotted. Past this frequency, the UI begins to lag. Possibly
consider moving parsing TX to a worker thread and batch updating the UI
state at set intervals instead of firing setState as data comes in.