perf: limit re-renders for realtime data
Because handleRoverTX was updating the component state for IMU data,
which is transmitted at 4 Hz, React would attempt to re-render the UI
consistently, leading to poor performance and dropped messages. Now,
when IMU data is recieved, it is saved to a class variable rather than
triggering a state update. A timer, activated when a Bluetooth device is
connected, will then update the component state with the contents of the
class variable via deep copying every 500 seconds. This leads to only
two re-renders due to IMU data per second rather than 12. This also
allows increasing the transmission frequency of IMU data to 10 Hz
without crashing the app.