feat: add setting for Wake Lock
When the user is driving (mode 0x02) or logging data, it can be helpful
to keep the screen on. This commit adds a setting on devices that
supports the Screen Wake Lock API to keep the screen on under these
conditions. The setting is enabled by default.
The wake lock will be released when the conditions are no longer met or
if the rover is disconnected.