feat: react to voltage changes
This commit changes the transmission format for voltages from string to
float. If the voltage is below 13.2 V, the status box will turn red and
the user will not be able to enable remote control.
The Arduino should also disable the motor, however that is not
implemented in this commit.

Misc. language changes have been made for clarity.