fix: workaround ArduinoBLE perf degradation
As a workaround for the performance issues described by Issue 9,
https://github.com/angelolu/RoverDriver/issues/9, this commit asks the
SoC to perform a system reset after at the end of the disconnect
callback when compiling for the Arduino Nano 33 BLE. This should not
impact peripherals as motor controls and the relay are updated before
the reset, however this may cause errors on the motor controller that
were cleared to also be ignored by the Arduino. This is not a priority
as the operator is likely to power cycle the system if any
irregularities are observed, at which point the motor controllers
will report their existing faults. As well, we are not currently
clearing errors that halt the motors.

With this change no performance degradations have been observed after
many connection/disconnection cycles. A minimal delay is introduced
between disconnecting and the device being available to reconnect
(less than one second).