feat: improve speed control on touch
The + and - buttons for increasing and decreasing speed now send
keycodes insead of the target speed for faster response times.
Additionally, the buttons now use the same events as the D-pad for
better response on touch devices. For example, the operator can now use
the D=pad and change speed simultaneously on multi-touch devices.