feat: basic motor controller status
This commit adds the basic functionality required to show if the motor
controllers are online and voltage, current, and cycle state. The
current controller states are stored in a class variable/object and a
timer will update the react state every 500 ms (same mechanism as the
IMU charts) to limit the number of state changes.

Errors, other than controller offline are not currently parsed,
although they are recieved. Further commits will add this, as well as
logging functionality