feat: log motor status
Motor status includes controller states and feedback from the motor
encoders. If the user requests that motor status be logged, on each
firing of the logging timer the current states of each motor controller
class variable (ex. _motorControllerFR) is saved. When the log is
exported, a new function, generateMotorLog, is used to flatten and
build the output object containing state information at each time
interval.

This commit will export current, duty cycle target, duty cycle and
feedback. Other info in the motor controller class variable (ex.
voltage, errors) is saved but not exported as they are not expected to
be useful to the end user.