fix: use browser-fs-access for downloads
This commit changes the log file export/download logic to use
browser-fs-access. On newer browsers this shows a save dialog where the
user can input a name and location for the file to be saved. The library
will fall back to a file download if the File API is not available.

This change addresses the issue on mobile devices where if the PWA was
installed and the app is running in a standalone window that log files
may not download (and show no error) if a file with the same filename
existed due to the browser blocking multiple downloads and/or not
showing the accept prompt.

closes #10